% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_c_and_g_strands.R
\name{convert_bismarck_cytosine_reports_to_bedgraphs}
\alias{convert_bismarck_cytosine_reports_to_bedgraphs}
\title{Convert Bismark genome-wide cytosine reports to BedGraph files with methylation values for each CpG site.}
\usage{
convert_bismarck_cytosine_reports_to_bedgraphs(
  cytosine_report_files,
  files_0_based = FALSE,
  genome,
  require_both_strands = TRUE,
  output_directory,
  n_parallel_files = 1,
  decimal_places = 2
)
}
\arguments{
\item{cytosine_report_files}{Paths to genome-wide cytosine reports output by bismark_methylation_extractor --cytosine_report.}

\item{files_0_based}{A logical value indicating if files are 0-based. Default is FALSE.}

\item{genome}{A BSgenome object or the name of a BSgenome package from which to extract CpG sites.}

\item{require_both_strands}{A logical value indicating whether to remove CpG sites in a file
without reads coming from both strands. Default is TRUE.}

\item{output_directory}{Directory to save output bedGraphs. Directory is created if it doesn't already exist.}

\item{n_parallel_files}{Number of files to process in parallel. Default is 1.}

\item{decimal_places}{Number of decimal places to round methylation values to. Default value is 2.}
}
\value{
A vector with the paths to the created bedGraph files.
}
\description{
Convert Bismark genome-wide cytosine reports to BedGraph files with methylation values for each CpG site.
}
