% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_methylation.R
\name{diff_meth_methylsig}
\alias{diff_meth_methylsig}
\title{Test differential methylation in predefined genomic regions using methylSig}
\usage{
diff_meth_methylsig(
  meth_rse,
  genomic_regions,
  meth_reads_assay = "meth_reads",
  coverage_assay = "cov",
  max_sites_per_chunk = NULL,
  na.rm = TRUE,
  group_column,
  case,
  control,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{meth_rse}{A RangedSummarizedExperiment with methylation values.}

\item{genomic_regions}{GRanges object with regions to get methylated and unmethylated counts for.}

\item{meth_reads_assay}{The assay of meth_rse which corresponds to the read counts for methylated reads.
Can be either a numeric index or the name of the assay. Default is "meth_reads".}

\item{coverage_assay}{The assay of meth_rse which corresponds to the total read counts for methylated and unmethylated reads.}

\item{max_sites_per_chunk}{The approximate maximum number of methylation sites to try to load into memory at once.
The actual number loaded may vary depending on the number of methylation sites overlapping each region,
but so long as the size of any individual regions is not enormous (>= several MB), it should vary only very slightly.
Some experimentation may be needed to choose an optimal value as low values will result in increased running time,
while high values will result in a large memory footprint without much improvement in running time.
Default is floor(62500000/ncol(meth_rse)), resulting in each chunk requiring approximately 500 MB of RAM.}

\item{na.rm}{TRUE or FALSE indicating whether to remove NA values when calculating summaries. Default value is TRUE.}

\item{group_column}{Name of column in colData(meth_rse) indicating the groups to be compared.}

\item{case}{Which level of group_column corresponds to the group of interest.}

\item{control}{Which level of group_column corresponds to the control group.}

\item{BPPARAM}{A BiocParallelParam object. Defaults to \code{BiocParallel::bpparam()}.}
}
\value{
A GRanges object with the results of methylSig::diff_methylsig().
}
\description{
First calculates the total number of methylated reads and the total number of reads
for all CpG sites in each region in each sample before testing differential methylation using
the beta-binomial test approach from methylSig.
}
