% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpg_summaries.R
\name{.summarize_cpgs_in_chunk}
\alias{.summarize_cpgs_in_chunk}
\title{Summarize methylation values for CpG sites in regions within a chunk}
\usage{
.summarize_cpgs_in_chunk(
  chunk_regions,
  meth_rse,
  genomic_regions,
  assay,
  summary_function_list,
  na.rm
)
}
\arguments{
\item{chunk_regions}{Chunk with genomic regions of interest.}

\item{meth_rse}{A RangedSummarizedExperiment with methylation values.}

\item{assay}{The assay from meth_rse to extract values from.
Should be either an index or the name of an assay.}

\item{summary_function_list}{A named list of functions to perform.}

\item{na.rm}{TRUE or FALSE indicating whether to remove NA values when calculating summaries.}
}
\value{
A list of data.frames with summaries of the methylation of CpG sites overlapping regions in the supplied chunk.
}
\description{
Summarize methylation values for CpG sites in regions within a chunk
}
