% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_c_and_g_strands.R
\name{extract_cpgs_from_bsgenome}
\alias{extract_cpgs_from_bsgenome}
\title{Create a GRanges with CpG sites from a BSgenome.}
\usage{
extract_cpgs_from_bsgenome(genome, standard_sequences_only = TRUE)
}
\arguments{
\item{genome}{A BSgenome object or the name of a BSgenome package.}

\item{standard_sequences_only}{TRUE or FALSE indicating whether to only return sites
on standard sequences (those without "-" in their names). Default is TRUE.}
}
\value{
A GRanges object CpG sites for the indicated genome.
}
\description{
Create a GRanges with CpG sites from a BSgenome.
}
\examples{
# Get human CpG sites for hg38 genome build
hg38_cpgs <- methodical::extractMethSitesFromGenome("BSgenome.Hsapiens.UCSC.hg38")

# Find CHG sites in Arabidopsis thaliana
arabidopsis_cphpgs <- methodical::extractMethSitesFromGenome("BSgenome.Athaliana.TAIR.TAIR9", pattern = "CHG")
}
