% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpg_summaries.R
\name{summarize_cpgs_in_regions}
\alias{summarize_cpgs_in_regions}
\title{Summarize methylation values for CpG sites in regions}
\usage{
summarize_cpgs_in_regions(
  meth_rse,
  genomic_regions,
  assay = 1,
  summary_function_list,
  na.rm = TRUE,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{meth_rse}{A RangedSummarizedExperiment with methylation values.}

\item{genomic_regions}{GRanges object with regions to summarize methylation values for.}

\item{assay}{The assay from meth_rse to extract values from.
Should be either an index or the name of an assay. Default is the first assay.}

\item{summary_function_list}{A named list of functions to perform.
Should be one of the row or column summary functions from matrixStats}

\item{na.rm}{TRUE or FALSE indicating whether to remove NA values when calculating summaries. Default is TRUE.}

\item{BPPARAM}{A BiocParallelParam object. Defaults to \code{BiocParallel::bpparam()}.}
}
\value{
A list of data.frames with summaries of the methylation of CpG sites overlapping each region in genomic_regions.
}
\description{
Summarize methylation values for CpG sites in regions
}
